+++ apps/astarte_data_updater_plant/lib/astarte_data_updater_plant_web/metrics.ex
@@ -19,5 +19,6 @@
 defmodule Astarte.DataUpdaterPlantWeb.Metrics do
   def setup do
     Astarte.DataUpdaterPlantWeb.Metrics.PrometheusExporter.setup()
+    Astarte.DataUpdaterPlantWeb.Metrics.DataUpdaterPlantInstrumenter.setup()
   end
 end