+++ apps/astarte_housekeeping/lib/astarte_housekeeping_web/metrics/supervisor.ex
@@ -29,7 +29,7 @@ defmodule Astarte.HousekeepingWeb.Metrics.Supervisor do
 
     # TODO: make the port configurable when we switch to Elixir native releases
     children = [
-      {Plug.Cowboy, scheme: :http, plug: Astarte.HousekeepingWeb.Router, options: [port: 4008]}
+      {Plug.Cowboy, scheme: :http, plug: Astarte.HousekeepingWeb.Router, options: [port: 4000]}
     ]
 
     Supervisor.init(children, strategy: :one_for_one)